requirements.txt
flask
gevent
flask_cors
Pillow
hurry.filesize
python-dotenv
APScheduler 